Resolve error with multi-digit operands
Resolve calculator error when using multi-digit operands. Previously,
the program did not function as intended because of whitespace added
behind every operand. (E.g., clicking 2 twice resulted in the text
content being "2 2 " instead of "22". The program uses the first 3 items
for performing calculation operations. In this case, the program
interpreted the input to be firstOperand = 2, operator = " ",
secondOperand = 2. " " is not a valid operator in our program, and hence
the program threw an error.)