Clear previous result upon new input
Previously, pressing a new digit after a result was displayed would
append the digit to the end of the already-evaluated expression. This
led to undesired behavior such as turning the display from "4 + 2" to
"4 + 22" if the user clicked the "2" button after "4 + 2" had been
evaluated. With this commit, if a result is currently being displayed
and the user presses a new digit, the display is cleared and a new
calculation is started with that new digit.